Refactor model definitions and relocate enums
Moved NutrientData struct from NutritionModels.swift to NutritionProgressSection.swift. Relocated TimeBasedCategory enum from NutritionModels.swift to TodayFoodLogView.swift. Deleted NutritionModels.swift as its contents were distributed to relevant component files. Also fixed a minor display issue in TodayFoodLogView.
